+++ ui/shared/components/panel/search/constants.js
@@ -1,8 +1,3 @@
-import { Form } from 'semantic-ui-react'
-import styled from 'styled-components'
-
-import { RadioGroup, InlineToggle } from 'shared/components/form/Inputs'
-
 export const THIS_CALLSET_FREQUENCY = 'callset'
 export const SV_CALLSET_FREQUENCY = 'sv_callset'
 export const TOPMED_FREQUENCY = 'topmed'
@@ -60,113 +55,3 @@ export const SV_FREQUENCIES = [
 ]
 
 export const FREQUENCIES = [...SNP_FREQUENCIES, ...MITO_FREQUENCIES, ...SV_FREQUENCIES]
-
-export const SNP_QUALITY_FILTER_FIELDS = [
-  {
-    name: 'affected_only',
-    label: 'Affected Only',
-    labelHelp: 'Only apply quality filters to affected individuals',
-    control: InlineToggle,
-    color: 'grey',
-    width: 6,
-  },
-  {
-    name: 'vcf_filter',
-    label: 'Filter Value',
-    labelHelp: 'Either show only variants that PASSed variant quality filters applied when the dataset was processed (typically VQSR or Hard Filters), or show all variants',
-    control: RadioGroup,
-    options: [{ value: '', text: 'Show All Variants' }, { value: 'pass', text: 'Pass Variants Only' }],
-    margin: '1em 2em',
-    widths: 'equal',
-  },
-  {
-    name: 'min_gq',
-    label: 'Genotype Quality',
-    labelHelp: 'Genotype Quality (GQ) is a statistical measure of confidence in the genotype call (eg. hom. or het) based on the read data',
-    min: 0,
-    max: 100,
-    step: 5,
-  },
-  {
-    name: 'min_ab',
-    label: 'Allele Balance',
-    labelHelp: 'The allele balance represents the percentage of reads that support the alt allele out of the total number of sequencing reads overlapping a variant. Use this filter to set a minimum percentage for the allele balance in heterozygous individuals.',
-    min: 0,
-    max: 50,
-    step: 5,
-  },
-]
-
-const DividedFormField = styled(Form.Field)`
-  border-left: solid grey 1px;
-`
-
-export const MITO_QUALITY_FILTER_FIELDS = [
-  {
-    name: 'min_hl',
-    label: 'Heteroplasmy level',
-    labelHelp: 'Heteroplasmy level (HL) is the percentage of the alt alleles out of all alleles.',
-    min: 0,
-    max: 50,
-    step: 5,
-    component: DividedFormField,
-  },
-]
-
-export const SV_QUALITY_FILTER_FIELDS = [
-  {
-    name: 'min_qs',
-    label: 'WES SV Quality Score',
-    labelHelp: 'The quality score (QS) represents the quality of a Structural Variant call. Recommended SV-QS cutoffs for filtering: duplication >= 50, deletion >= 100, homozygous deletion >= 400.',
-    min: 0,
-    max: 1000,
-    step: 10,
-    component: DividedFormField,
-  },
-  {
-    name: 'min_gq_sv',
-    label: 'WGS SV Genotype Quality',
-    labelHelp: 'The genotype quality (GQ) represents the quality of a Structural Variant call. Recommended SV-GQ cutoffs for filtering: > 10.',
-    min: 0,
-    max: 100,
-    step: 5,
-  },
-]
-
-export const QUALITY_FILTER_FIELDS = [
-  ...SNP_QUALITY_FILTER_FIELDS,
-  ...MITO_QUALITY_FILTER_FIELDS,
-  ...SV_QUALITY_FILTER_FIELDS,
-]
-
-export const ALL_QUALITY_FILTER = {
-  text: 'All Variants',
-  value: {
-    vcf_filter: null,
-    min_gq: 0,
-    min_ab: 0,
-    min_qs: 0,
-  },
-}
-
-export const QUALITY_FILTER_OPTIONS = [
-  ALL_QUALITY_FILTER,
-  {
-    text: 'High Quality',
-    value: {
-      vcf_filter: 'pass',
-      min_gq: 20,
-      min_ab: 25,
-      min_qs: 100,
-    },
-  },
-  {
-    text: 'All Passing Variants',
-    value: {
-      vcf_filter: 'pass',
-      min_gq: 0,
-      min_ab: 0,
-      min_qs: 10,
-    },
-  },
-]
